+++ src/app/components/file-browser/file-entry-charts.js
@@ -20,6 +20,7 @@ import {
   dirSizeStatsTimeSeriesNameGenerators as timeSeriesNameGenerators,
 } from 'oneprovider-gui/models/file';
 import { hashSettled, hash as hashFulfilled, all as allFulfilled } from 'rsvp';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
 const mixins = [
   I18n,
@@ -82,6 +83,11 @@ export default Component.extend(...mixins, {
    * @type {Looper}
    */
   latestDirSizeStatsValuesUpdater: undefined,
+
+  /**
+   * @type {Boolean}
+   */
+  dirStatsNotReady: false,
 
   /**
    * @type {ComputedProperty<string>}
@@ -750,7 +756,17 @@ export default Component.extend(...mixins, {
    * @returns {Promise<DirCurrentSizeStats>}
    */
   async fetchLatestDirSizeStatsValues() {
-    return this.fileManager.getDirCurrentSizeStats(this.fileId);
+    try {
+      const result = await this.fileManager.getDirCurrentSizeStats(this.fileId);
+      safeExec(this, () => this.set('dirStatsNotReady', false));
+      return result;
+    } catch (error) {
+      const dirStatsNotReady = error?.id === 'dirStatsNotReady';
+      safeExec(this, () => this.set('dirStatsNotReady', dirStatsNotReady));
+      if (!dirStatsNotReady) {
+        throw error;
+      }
+    }
   },
 
   actions: {